+++ docs/CHANGE_LOG.md
@@ -6,8 +6,11 @@
 - BREAKING: All binding options that affect the views and how they look are now split up into categories under a new "views" area.
 
 #### **Binding Options:**
-- BREAKING: Renamed the binding option "mapRangeColors" to "colorRanges" (more consistent with the project), and renamed documentation.
+- BREAKING: Renamed the binding option "mapRangeColors" to "colorRanges" (more consistent with the project), and updated documentation.
 - Added a new binding option called "views.statistics.showColorRangeLabels", which states if the chart X color range labels should be shown (defaults to true).
+
+#### **Binding Options - Custom Triggers:**
+- BREAKING: Renamed the binding option custom trigger "onMapRangeTypeToggle" to "onColorRangeTypeToggle" (more consistent with the project), and updated documentation.
 
 <br>